+++ lib/flutter_device_config.dart
@@ -4,6 +4,7 @@ export 'src/body.dart';
 export 'src/config.dart';
 export 'src/constants.dart';
 export 'src/device.dart';
+export 'src/dimen.dart';
 export 'src/extension.dart';
 export 'src/layout.dart';
 export 'src/type.dart';