+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 1.0.3
+
+* DimensionSize support [dp, dx, dy]
+* DimensionScaleMode support [maximum, minimum, both, none]
+
+## 1.0.2
+
+* Assumed device dimension supports [dp, dx, dy]
+
 ## 1.0.1
 
 * initial release.